Extract organization breadcrumb stuff to its own separate helper
